feat: add prompt saving as message

+++ backend/src/main/java/ru/hh/summarizer/dao/ChatRepository.java
@@ -3,5 +3,5 @@ package ru.hh.summarizer.dao;
 import org.springframework.data.repository.CrudRepository;
 import ru.hh.summarizer.entity.Chat;
 
-public interface ChatRepository extends CrudRepository<Chat, Integer> {
+public interface ChatRepository extends CrudRepository<Chat, Long> {
 }
